automatically backed up by learn

+++ lib/create.sql
@@ -3,8 +3,8 @@ id INTEGER PRIMARY KEY,
 title TEXT,
 category TEXT,
 funding_goal INTEGER,
-start_date INTEGER,
-end_date INTEGER);
+start_date REAL,
+end_date REAL);
 
 CREATE TABLE users (
 id INTEGER PRIMARY KEY,
